+++ src/app/features/property/property.module.ts
@@ -6,18 +6,40 @@ import {CreatePropertyComponent} from "./create-property/create-property.compone
 import {RouterLink} from "@angular/router";
 import {UpdatePropertyComponent} from './update-property/update-property.component';
 import {PropertyDetailComponent} from './property-detail/property-detail.component';
+import {MatTableModule} from "@angular/material/table";
+import {MatFormFieldModule} from "@angular/material/form-field";
+import {MatInputModule} from "@angular/material/input";
+import {MatIconModule} from "@angular/material/icon";
+import {PropertyFormComponent} from './property-form/property-form.component';
+import {MatStepperModule} from "@angular/material/stepper";
+import {ReactiveFormsModule} from "@angular/forms";
+import {MatDividerModule} from "@angular/material/divider";
+import {MatRadioModule} from "@angular/material/radio";
+import {ComponentsModule} from "../../common/components/components.module";
+import {MatCardModule} from "@angular/material/card";
 
 @NgModule({
   declarations: [
     PropertiesDashboardComponent,
     CreatePropertyComponent,
     UpdatePropertyComponent,
-    PropertyDetailComponent
+    PropertyDetailComponent,
+    PropertyFormComponent
   ],
   imports: [
     CommonModule,
     MatButtonModule,
-    RouterLink
+    RouterLink,
+    MatTableModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatIconModule,
+    MatStepperModule,
+    ReactiveFormsModule,
+    MatDividerModule,
+    MatRadioModule,
+    ComponentsModule,
+    MatCardModule
   ]
 })
 export class PropertyModule {